moves suspend function to coroutine
