+++ purchases/src/main/kotlin/com/revenuecat/purchases/coroutines.kt
@@ -1,0 +1,27 @@
+package com.revenuecat.purchases
+
+import com.revenuecat.purchases.interfaces.ReceiveCustomerInfoCallback
+import com.revenuecat.purchases.utils.Result
+import kotlin.coroutines.resume
+import kotlin.coroutines.suspendCoroutine
+
+/**
+ * Get latest available purchaser info.
+ * Coroutine friendly version of [getCustomerInfo].
+ *
+ * @return [Result] type containing either the [CustomerInfo] or the [PurchasesError]
+ */
+suspend fun Purchases.getCustomerInfo(): Result<CustomerInfo, PurchasesError> {
+    return suspendCoroutine { continuation ->
+        val receiveCustomerInfoCallback = object : ReceiveCustomerInfoCallback {
+            override fun onReceived(customerInfo: CustomerInfo) {
+                continuation.resume(Result.Success(customerInfo))
+            }
+
+            override fun onError(error: PurchasesError) {
+                continuation.resume(Result.Error(error))
+            }
+        }
+        getCustomerInfo(receiveCustomerInfoCallback)
+    }
+}
